Make Settings Panel tabular
Refactored and organized the UI files into appropriate folders
TODO: Add all the new settings
