+++ mp/game/momentum/resource/ui/SettingsPanel_Base.res
@@ -1,0 +1,369 @@
+"resource/ui/SettingsPanel_Base.res"
+{
+	"CMomentumSettingsPanel"
+	{
+		"ControlName"		"CMomentumSettingsPanel"
+		"fieldName"		"CMomentumSettingsPanel"
+		"xpos"		"304"
+		"ypos"		"144"
+		"wide"		"266"
+		"tall"		"266"
+		"autoResize"		"1"
+		"pinCorner"		"0"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"1"
+		"enabled"		"1"
+		"tabPosition"		"0"
+		"settitlebarvisible"		"1"
+		"title"		"#MOM_Settings"
+	}
+    "Sheet"
+    {
+        "ControlName" "PropertySheet"
+        "fieldName" "Sheet"
+    }
+    
+    "OKButton"
+    {
+       "ControlName" "Button"
+       "fieldName" "OKButton"
+       "auto_wide_tocontents" "1"
+       "textAlignment"		"center"
+       //"font" "DefaultSmall"
+    }
+    
+    "CancelButton"
+    {
+       "ControlName" "Button"
+       "fieldName" "CancelButton"
+       "auto_wide_tocontents" "1"
+       "textAlignment"		"center"
+       //"font" "DefaultSmall"
+    }
+    
+    "ApplyButton"
+    {
+       "ControlName" "Button"
+       "fieldName" "ApplyButton"
+       "auto_wide_tocontents" "1"
+       "textAlignment"		"center"
+       //"font" "DefaultSmall"
+    }
+    
+    
+	"frame_topGrip"
+	{
+		"ControlName"		"Panel"
+		"fieldName"		"frame_topGrip"
+		"xpos"		"4"
+		"ypos"		"0"
+		"wide"		"258"
+		"tall"		"2"
+		"autoResize"		"0"
+		"pinCorner"		"0"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"1"
+		"enabled"		"1"
+		"tabPosition"		"0"
+	}
+	"frame_bottomGrip"
+	{
+		"ControlName"		"Panel"
+		"fieldName"		"frame_bottomGrip"
+		"xpos"		"4"
+		"ypos"		"264"
+		"wide"		"252"
+		"tall"		"2"
+		"autoResize"		"0"
+		"pinCorner"		"0"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"1"
+		"enabled"		"1"
+		"tabPosition"		"0"
+	}
+	"frame_leftGrip"
+	{
+		"ControlName"		"Panel"
+		"fieldName"		"frame_leftGrip"
+		"xpos"		"0"
+		"ypos"		"4"
+		"wide"		"2"
+		"tall"		"258"
+		"autoResize"		"0"
+		"pinCorner"		"0"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"1"
+		"enabled"		"1"
+		"tabPosition"		"0"
+	}
+	"frame_rightGrip"
+	{
+		"ControlName"		"Panel"
+		"fieldName"		"frame_rightGrip"
+		"xpos"		"264"
+		"ypos"		"4"
+		"wide"		"2"
+		"tall"		"252"
+		"autoResize"		"0"
+		"pinCorner"		"0"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"1"
+		"enabled"		"1"
+		"tabPosition"		"0"
+	}
+	"frame_tlGrip"
+	{
+		"ControlName"		"Panel"
+		"fieldName"		"frame_tlGrip"
+		"xpos"		"0"
+		"ypos"		"0"
+		"wide"		"4"
+		"tall"		"4"
+		"autoResize"		"0"
+		"pinCorner"		"0"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"1"
+		"enabled"		"1"
+		"tabPosition"		"0"
+	}
+	"frame_trGrip"
+	{
+		"ControlName"		"Panel"
+		"fieldName"		"frame_trGrip"
+		"xpos"		"262"
+		"ypos"		"0"
+		"wide"		"4"
+		"tall"		"4"
+		"autoResize"		"0"
+		"pinCorner"		"0"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"1"
+		"enabled"		"1"
+		"tabPosition"		"0"
+	}
+	"frame_blGrip"
+	{
+		"ControlName"		"Panel"
+		"fieldName"		"frame_blGrip"
+		"xpos"		"0"
+		"ypos"		"262"
+		"wide"		"4"
+		"tall"		"4"
+		"autoResize"		"0"
+		"pinCorner"		"0"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"1"
+		"enabled"		"1"
+		"tabPosition"		"0"
+	}
+	"frame_brGrip"
+	{
+		"ControlName"		"Panel"
+		"fieldName"		"frame_brGrip"
+		"xpos"		"257"
+		"ypos"		"257"
+		"wide"		"9"
+		"tall"		"9"
+		"autoResize"		"0"
+		"pinCorner"		"0"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"1"
+		"enabled"		"1"
+		"tabPosition"		"0"
+	}
+	"frame_caption"
+	{
+		"ControlName"		"Panel"
+		"fieldName"		"frame_caption"
+		"xpos"		"0"
+		"ypos"		"0"
+		"wide"		"261"
+		"tall"		"12"
+		"autoResize"		"0"
+		"pinCorner"		"0"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"1"
+		"enabled"		"1"
+		"tabPosition"		"0"
+	}
+	"frame_minimize"
+	{
+		"ControlName"		"Button"
+		"fieldName"		"frame_minimize"
+		"xpos"		"0"
+		"ypos"		"0"
+		"wide"		"0"
+		"tall"		"0"
+		"autoResize"		"0"
+		"pinCorner"		"0"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"0"
+		"enabled"		"1"
+		"tabPosition"		"0"
+		"labelText"		"0"
+		"textAlignment"		"north-west"
+		"dulltext"		"0"
+		"brighttext"		"0"
+		"wrap"		"0"
+		"centerwrap"		"0"
+		"textinsetx"		"2"
+		"textinsety"		"1"
+		"auto_wide_tocontents"		"1"
+		"use_proportional_insets"		"0"
+		"Default"		"0"
+	}
+	"frame_maximize"
+	{
+		"ControlName"		"Button"
+		"fieldName"		"frame_maximize"
+		"xpos"		"0"
+		"ypos"		"0"
+		"wide"		"0"
+		"tall"		"0"
+		"autoResize"		"0"
+		"pinCorner"		"0"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"0"
+		"enabled"		"1"
+		"tabPosition"		"0"
+		"labelText"		"1"
+		"textAlignment"		"north-west"
+		"dulltext"		"0"
+		"brighttext"		"0"
+		"wrap"		"0"
+		"centerwrap"		"0"
+		"textinsetx"		"2"
+		"textinsety"		"1"
+		"auto_wide_tocontents"		"1"
+		"use_proportional_insets"		"0"
+		"Default"		"0"
+	}
+	"frame_mintosystray"
+	{
+		"ControlName"		"Button"
+		"fieldName"		"frame_mintosystray"
+		"xpos"		"0"
+		"ypos"		"0"
+		"wide"		"0"
+		"tall"		"0"
+		"autoResize"		"0"
+		"pinCorner"		"0"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"0"
+		"enabled"		"1"
+		"tabPosition"		"0"
+		"labelText"		"o"
+		"textAlignment"		"north-west"
+		"dulltext"		"0"
+		"brighttext"		"0"
+		"wrap"		"0"
+		"centerwrap"		"0"
+		"textinsetx"		"2"
+		"textinsety"		"1"
+		"auto_wide_tocontents"		"1"
+		"use_proportional_insets"		"0"
+		"command"		"MinimizeToSysTray"
+		"Default"		"0"
+	}
+	"frame_close"
+	{
+		"ControlName"		"Button"
+		"fieldName"		"frame_close"
+		"xpos"		"242"
+		"ypos"		"8"
+		"wide"		"17"
+		"tall"		"17"
+		"autoResize"		"0"
+		"pinCorner"		"0"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"1"
+		"enabled"		"1"
+		"tabPosition"		"0"
+		"labelText"		"r"
+		"textAlignment"		"north-west"
+		"dulltext"		"0"
+		"brighttext"		"0"
+		"wrap"		"0"
+		"centerwrap"		"0"
+		"textinsetx"		"4"
+		"textinsety"		"2"
+		"auto_wide_tocontents"		"1"
+		"use_proportional_insets"		"0"
+		"Default"		"0"
+	}
+	"frame_menu"
+	{
+		"ControlName"		"FrameSystemButton"
+		"fieldName"		"frame_menu"
+		"xpos"		"3"
+		"ypos"		"4"
+		"wide"		"9"
+		"tall"		"9"
+		"autoResize"		"0"
+		"pinCorner"		"0"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"1"
+		"enabled"		"1"
+		"tabPosition"		"0"
+		"textAlignment"		"west"
+		"dulltext"		"0"
+		"brighttext"		"0"
+		"wrap"		"0"
+		"centerwrap"		"0"
+		"textinsetx"		"0"
+		"textinsety"		"0"
+		"auto_wide_tocontents"		"1"
+		"use_proportional_insets"		"0"
+		"Default"		"0"
+	}
+	
+	"BuildModeDialog"
+	{
+		"ControlName"		"BuildModeDialog"
+		"fieldName"		"BuildModeDialog"
+		"xpos"		"33"
+		"ypos"		"19"
+		"wide"		"160"
+		"tall"		"224"
+		"autoResize"		"0"
+		"pinCorner"		"0"
+		"RoundedCorners"		"15"
+		"pin_corner_to_sibling"		"0"
+		"pin_to_sibling_corner"		"0"
+		"visible"		"1"
+		"enabled"		"1"
+		"tabPosition"		"0"
+		"settitlebarvisible"		"1"
+		"title"		"#Frame_Untitled"
+	}
+}
