+++ mp/src/game/client/client_momentum.vpc
@@ -13,7 +13,7 @@ $Configuration
 {
     $Compiler
     {
-        $AdditionalIncludeDirectories       "$BASE;momentum\ui\MapSelection,.\momentum\ui,.\momentum;$SRCDIR\game\shared\momentum;$SRCDIR\thirdparty\gason\src"
+        $AdditionalIncludeDirectories       "$BASE;momentum\ui\HUD,.\momentum\ui\SettingsPanel,.\momentum\ui\MapSelection,.\momentum\ui,.\momentum;$SRCDIR\game\shared\momentum;$SRCDIR\thirdparty\gason\src"
         $PreprocessorDefinitions            "$BASE;SOURCE_2013"
     }
 }
@@ -62,30 +62,42 @@ $Project "Client (Momentum)"
                     }
                 }
                 
-                $File "momentum\ui\MomentumSettingsDialog.cpp"
-                $File "momentum\ui\IMomentumSettingsPanel.h"
-                $File   "momentum\ui\hud_speedometer.cpp"
-                $File   "momentum\ui\hud_strafesync.cpp"
-                $File   "momentum\ui\hud_timer.cpp"
-                $File   "momentum\ui\hud_cp_menu.cpp"
-                $File   "momentum\ui\hud_menu_static.h"
-                $File   "momentum\ui\hud_menu_static.cpp"
-                $File   "momentum\ui\hud_fillablebar.h"
+                $Folder "SettingsPanel"
+                {
+                    $File "momentum\ui\SettingsPanel\MomentumSettingsDialog.cpp"
+                    $File "momentum\ui\SettingsPanel\IMomentumSettingsPanel.h"
+                    $File "momentum\ui\SettingsPanel\SettingsPage.h"
+                    $File "momentum\ui\SettingsPanel\HudSettingsPage.h"
+                }
+                
+                $Folder "HUD"
+                {
+                    $File "momentum\ui\HUD\hud_versioninfo.cpp"
+                    $File "momentum\ui\HUD\hud_mapfinished.cpp"
+                    $File "momentum\ui\HUD\hud_mapfinished.h"
+                    $File "momentum\ui\HUD\hud_keypress.cpp"
+                    $File "momentum\ui\HUD\hud_comparisons.cpp"
+                    $File "momentum\ui\HUD\hud_comparisons.h"
+                    $File "momentum\ui\HUD\hud_mapinfo.cpp"
+                    $File "momentum\ui\HUD\hud_speedometer.cpp"
+                    $File "momentum\ui\HUD\hud_strafesync.cpp"
+                    $File "momentum\ui\HUD\hud_timer.cpp"
+                    $File "momentum\ui\HUD\hud_cp_menu.cpp"
+                    $File "momentum\ui\HUD\hud_menu_static.h"
+                    $File "momentum\ui\HUD\hud_menu_static.cpp"
+                    $File "momentum\ui\HUD\hud_fillablebar.h"
+                    $File "hl2\hud_credits.cpp"
+                    $File "hud_chat.cpp"
+                    $File "hl2\hud_weaponselection.cpp"
+                }
+                
                 $File   "momentum\ui\ClientTimesDisplay.h"
                 $File   "momentum\ui\ClientTimesDisplay.cpp"
+                
                 $File   "momentum\ui\IVersionWarnPanel.h"
                 $File   "momentum\ui\VersionWarnPanel.cpp"
                 $File   "momentum\ui\VersionWarnPanel.h"
-                $File   "momentum\ui\hud_versioninfo.cpp"
-                $File "momentum\ui\hud_mapfinished.cpp"
-                $File "momentum\ui\hud_mapfinished.h"
-                $File "momentum\ui\hud_keypress.cpp"
-                $File "hl2\hud_credits.cpp"
-                $File "hl2\hud_weaponselection.cpp"
-                $File "hud_chat.cpp"
-                $File "momentum\ui\hud_comparisons.cpp"
-                $File "momentum\ui\hud_comparisons.h"
-                $File "momentum\ui\hud_mapinfo.cpp"
+                
                 $File "momentum\ui\momSpectatorGUI.cpp"
                 $File "momentum\ui\momSpectatorGUI.h"
             }
