+++ mp/src/game/client/vgui_int.cpp
@@ -27,7 +27,7 @@
 //Momentum
 #include "momentum/ui/MapSelection/IMapSelector.h"
 #include "momentum/ui/IVersionWarnPanel.h"
-#include "momentum/ui/IMomentumSettingsPanel.h"
+#include "momentum/ui/SettingsPanel/IMomentumSettingsPanel.h"
 
 #ifdef SIXENSE
 #include "sixense/in_sixense.h"
